requirements.txt
RPi.GPIO
requests
configparser
